chore: move UI experiments to separate route
