+++ app/(tools)/write-emmc/page.tsx
@@ -1,3 +1,4 @@
-export default function SampleRoute() {
+'use client'
+export default function WriteEmmc() {
   return <div>Sample Route, to be populated later.</div>
 }
