+++ app/page.tsx
@@ -1,3 +1,4 @@
+'use client'
 export default function Home() {
-  return <div>Sample Route</div>
+  return <div>Sample Route, to be populated later.</div>
 }
